we now delete files when a result is deleted

+++ controllers/results.mjs
@@ -10,6 +10,9 @@ import * as commonTools from '../utils/common.mjs';
 
 import fs from 'fs';
 import path from 'path';
+import { fileURLToPath } from 'url';
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 resultsRouter.get('/', async (request, response, next) => {
   try {
@@ -37,14 +40,16 @@ resultsRouter.delete('/:id', async (request, response, next) => {
 
     // TODO: delete the files (snapshots or reports) for this result
     const toBeDeletedDocument = await Result.findById(objectId);
-    toBeDeletedDocument.outcome.files.forEach(file => {
-      if (file.endsWith("html")) {
-        fs.unlinkSync(path.resolve(__dirname, "../public/reports/" + file));
-      }
-      else if (file.endsWith("png")) {
-        fs.unlinkSync(path.resolve(__dirname, "../public/snapshots/" + file));
-      }
-    })
+    if (toBeDeletedDocument.outcome.files) {
+      toBeDeletedDocument.outcome.files.forEach(file => {
+        if (file.endsWith("html")) {
+          fs.unlinkSync(path.resolve(__dirname, "../public/reports/" + file));
+        }
+        else if (file.endsWith("png")) {
+          fs.unlinkSync(path.resolve(__dirname, "../public/snapshots/" + file));
+        }
+      })
+    }
     const deletedDocument = await Result.findByIdAndDelete(objectId);
 
     if (deletedDocument) {
